Move class to src folder
